requirements.txt
numpy
pandas
scikit-learn==0.23.2
scikit-surprise
git+https://github.com/statisticianinstilettos/recmetrics.git
jupyter
fuzzywuzzy
python-Levenshtein